wip: file-recorder first implem

+++ buildAll.sh
@@ -6,10 +6,10 @@ if [ ! -d "$build_dir" ]; then
 fi
 
 if [ "$(ls -A "$build_dir")" ]; then
-    rm -r "$build_dir"/*
+    rm -r "${build_dir:?}/"*
 fi
 
-files=("sensors/pressureSensor" "sensors/tempSensor" "sensors/windSensor")
+files=("sensors/pressureSensor" "sensors/tempSensor" "sensors/windSensor" "file-recorder")
 
 for file in "${files[@]}"; do
     echo "Building $file"
